requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==37.1.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
pytest==8.3.5
pytest-django==4.11.1
pytest-factoryboy==2.7.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
